分离出pg和session的代码
